lsコマンドのファイルを分ける
